requirements.txt
alembic==1.7.7
aniso8601==9.0.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click @ file:///C:/ci/click_1646038595831/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
dnspython==2.2.1
email-validator==1.1.3
Flask @ file:///opt/conda/conda-bld/flask_1648041541647/work
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
imageio==2.17.0
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
Mako==1.2.0
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
numpy==1.22.3
Pillow==9.1.0
pycparser==2.21
pytz==2022.1
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.35
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
wincertstore==0.2
WTForms==3.0.1
